refactor: refactor move mouse functionality
