+++ src/types/mouse.ts
